+++ test/test06.bas
@@ -1,7 +1,7 @@
-sub abs(n): return n end
-sub f(a, b): return 0 end
-sub g(c): return 0: end
-sub h(): return 0 end
+sub abs(n) as int32: return n end
+sub f(a, b) as int32: return 0 end
+sub g(c) as int32: return 0: end
+sub h() as int32: return 0 end
 
 SUB main(a, b, c)
     PRINT 1 * 2 + 3 * 4 + abs(5 + 6 * 7)
