+++ test/test17.bas
@@ -1,20 +1,20 @@
-sub f1(a)
+sub f1(a) as int32
     return a
 end
 
-sub f2(a, b)
+sub f2(a, b) as int32
     return b
 end
 
-sub f3(a, b, c)
+sub f3(a, b, c) as int32
     return c
 end
 
-sub f4(a, b, c, d)
+sub f4(a, b, c, d) as int32
     return d
 end
 
-sub f5(a, b, c, d, e)
+sub f5(a, b, c, d, e) as int32
     return e
 end
