+++ test/test19.bas
@@ -1,4 +1,4 @@
-sub foo(a as int64)
+sub foo(a as int64) as int64
     return a
 end
